Move `RemoteFeedImageDataLoader` to production scope
